# api-document.yaml
openapi: 3.0.0
info:
  title: 'ゼミのホームページ'
  description: 'ラシキアゼミのホームページのapi'
  version: '1.0.0'
...
paths:
...
  /researches:
    get:
      summary: '卒業研究を全件取得'
      tags: 
        - 'research'
      responses:
        '200':
          description: '卒業研究のデータ'
          content:
            application/json:
              schema:
                type: 'object'
                properties:
                  researches:
                    type: 'array'
                    items:
                      $ref: '#/components/schemas/GetResearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: '卒業研究のデータを作成'
      description: 'ユーザがowner/adminである必要がある'
      tags: 
        - 'research'
      parameters: 
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        $ref: '#/components/requestBodies/UpdateResearch'
      responses:
        '200':
          description: '創業研究のデータ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetResearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    AccessToken:
      type: 'string'
...
    GetResearchResponse:
      type: 'object'
      properties:
        id:
          type: 'integer'
        title:
          type: 'string'
        author:
          type: 'string'
        file:
          type: 'string'
        comment:
          type: 'string'
        activation:
          type: 'integer'
    UpdateResearchRequest:
      type: 'object'
      properties:
        title:
          type: 'string'
        author:
          type: 'string'
        file:
          type: 'string'
        comment:
          type: 'string'
...
    ErrorResponse:
      type: 'object'
      properties:
        code:
          type: 'integer'
        message:
          type: 'string'
  parameters:
    SessionToken:
      name: '_l_semi_homepage_session'
      in: 'cookie'
      required: true
      schema:
        $ref: '#/components/schemas/AccessToken'
  requestBodies:
...
    UpdateResearch:
      description: '卒研発表の登録。新規なら全てのパラメータが必須。更新なら一つ以上'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateResearchRequest'
...
  responses:
    BadRequest:
      description: 'bad request. status code: 400'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 'unauthorized. status code: 401'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: 'internal server error. status code: 500'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...